+++ karma-cov.conf.js
@@ -1,8 +1,6 @@
 // Defines a test server for running jasmine unit tests
 // with coverage support.
 
-var path = require('path');
-
 /**
  * Return URL friendly browser string
  */
@@ -34,17 +32,17 @@ module.exports = function (config) {
   karma_config.specReporter = {suppressPassed: true, suppressSkipped: true};
   karma_config.coverageReporter = {
     reporters: [
-      {type: 'html', dir: 'dist/coverage/', subdir: subdir_name},
       {type: 'cobertura', dir: 'dist/cobertura/', file: 'coverage.xml', subdir: subdir_name},
       {type: 'json', dir: 'dist/coverage/json/', subdir: subdir_name},
-      {type: 'lcovonly', dir: 'lcov', subdir: subdir_name},
+      {type: 'lcovonly', dir: 'dist/coverage/lcov', subdir: subdir_name},
       {type: 'text'}
     ]
   };
-  karma_config.webpack.module.rules.unshift({
-    test: /\.js$/,
-    include: path.resolve('src/'),
-    use: ['istanbul-instrumenter-loader']
+  /* Alter our first webpack module rule which should just apply to src/*.js
+   * files. */
+  karma_config.webpack.module.rules[0].use.push({
+    loader: 'istanbul-instrumenter-loader',
+    options: {esModules: true}
   });
 
   config.set(karma_config);