+++ src/gl/tileLayer.js
@@ -37,7 +37,12 @@ var gl_tileLayer = function () {
     quad.lr = this.fromLocal(this.fromLevel({
       x: bounds.right - to.x, y: bounds.bottom - to.y
     }, level), 0);
-    quad.ul.z = quad.ll.z = quad.ur.z = quad.lr.z = level * 1e-5;
+    /* Make sure our level increments are within the clipbounds and ordered so
+     * that lower levels are farther away that higher levels. */
+    var clipbounds = m_this.map().camera().clipbounds;
+    var z = level * m_this._levelZIncrement;
+    z = clipbounds.far + (clipbounds.near - clipbounds.far) * z;
+    quad.ul.z = quad.ll.z = quad.ur.z = quad.lr.z = z;
     m_nextTileId += 1;
     quad.id = m_nextTileId;
     tile.quadId = quad.id;