+++ src/tileLayer.js
@@ -196,6 +196,11 @@ var tileLayer = function (options) {
       m_maxBounds = [],
       m_exited;
 
+  // Space tile levels this far apart in z-buffer space.  This value doesn't
+  // have any visual effect in parallel projections, but will slightly skew
+  // perspective projections.
+  this._levelZIncrement = 1e-5;
+
   // copy the options into a private variable
   this._options = $.extend(true, {}, options);