+++ CHANGELOG.md
@@ -2,10 +2,17 @@
 
 ## Unreleased
 
+### Features
+- Added an idle property to objects (#894)
+- Better handling and changing of camera clipbounds (#899)
+
+### Bug Fixes
+- Fixed an issue with overlapping, cropped tiles on old browsers (#901)
+- Fixed an issue where a `geo.gl.polygonFeature` could be updated after it was deleted (#913)
+
 ### Changes
-
 - Changed build process: optional dependencies are now included in the bundle by default (#890)
-
+- Transpile with Babel to support old browsers and new language features (#900)
 
 ## Version 0.17.0